[Playground] [Backend] removed extra methods and files

+++ playground/backend/internal/db/entity/id_meta.go
@@ -17,5 +17,5 @@ package entity
 
 type IDMeta struct {
 	Salt     string
-	IdLength int
+	IdLength int8
 }
